Renamed dir "views" to "pages". Added EP component el-menu and el-menu-item. Added navigation between pages in App component.
